+++ src/main/java/org/folio/entlinks/domain/repository/LinkingRulesRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.folio.entlinks.domain.entity.InstanceAuthorityLinkingRule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LinkingRulesRepository extends JpaRepository<InstanceAuthorityLinkingRule, String> {
+public interface LinkingRulesRepository extends JpaRepository<InstanceAuthorityLinkingRule, Integer> {
 
   List<InstanceAuthorityLinkingRule> findByAuthorityField(String authorityField);
 }