+++ src/main/java/org/folio/entlinks/service/links/InstanceAuthorityLinkingRulesService.java
@@ -7,8 +7,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.entlinks.domain.entity.InstanceAuthorityLinkingRule;
 import org.folio.entlinks.domain.repository.LinkingRulesRepository;
+import org.folio.entlinks.exception.LinkingRuleNotFoundException;
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Log4j2
 @Service
@@ -19,7 +23,7 @@ public class InstanceAuthorityLinkingRulesService {
 
   public List<InstanceAuthorityLinkingRule> getLinkingRules() {
     log.info("Loading linking rules");
-    return repository.findAll();
+    return repository.findAll(Sort.by("id").ascending());
   }
 
   @Cacheable(cacheNames = AUTHORITY_LINKING_RULES_CACHE,
@@ -29,4 +33,22 @@ public class InstanceAuthorityLinkingRulesService {
     return repository.findByAuthorityField(authorityField);
   }
 
+  public InstanceAuthorityLinkingRule getLinkingRule(Integer ruleId) {
+    log.info("Loading linking rule [ruleId: {}]", ruleId);
+    return repository.findById(ruleId)
+      .orElseThrow(() -> new LinkingRuleNotFoundException(ruleId));
+  }
+
+  @Transactional
+  @CacheEvict(cacheNames = AUTHORITY_LINKING_RULES_CACHE, allEntries = true)
+  public void patchLinkingRule(Integer ruleId, InstanceAuthorityLinkingRule linkingRulePatch) {
+    log.info("Patch linking rule [ruleId: {}, change: {}]", ruleId, linkingRulePatch);
+    var existedLinkingRule = repository.findById(ruleId)
+      .orElseThrow(() -> new LinkingRuleNotFoundException(ruleId));
+    // only autoLinkingEnabled flag is allowed to be updated right now
+    if (linkingRulePatch.getAutoLinkingEnabled() != null) {
+      existedLinkingRule.setAutoLinkingEnabled(linkingRulePatch.getAutoLinkingEnabled());
+    }
+    repository.save(existedLinkingRule);
+  }
 }